Moves all markup for the screen launcher into the screen launcher component
