+++ src/components/MenuComponent.js
@@ -3,50 +3,16 @@
 import React from 'react';
 import FloatingActionButton from 'material-ui/FloatingActionButton';
 import AvSkipPrevious from 'material-ui/svg-icons/av/skip-previous';
-import ContentLowPriority from 'material-ui/svg-icons/content/low-priority';
 import AvPause from 'material-ui/svg-icons/av/pause';
 import AvPlayArrow from 'material-ui/svg-icons/av/play-arrow';
 import AvSkipNext from 'material-ui/svg-icons/av/skip-next';
 import KeyBinding from 'react-keybinding-component';
-import FlatButton from 'material-ui/FlatButton';
-import Dialog from 'material-ui/Dialog';
-import ScreenLauncher from './ScreenLauncherComponent';
+import ScreenLauncherComponent from './ScreenLauncherComponent';
 import SettingsComponent from './SettingsComponent';
-import { defineMessages, injectIntl } from 'react-intl';
 
 require('styles//Menu.scss');
 
-const messages = defineMessages({
-  title: {
-    id: 'screenlauncher.title',
-    defaultMessage: 'Go to screen'
-  },
-  closeButton: {
-    id: 'screenlauncher.close',
-    defaultMessage: 'Close'
-  }
-});
-
 class MenuComponent extends React.Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      screenLauncherModalOpen: false
-    };
-  }
-
-  handleOpen = () => {
-    this.setState({
-      screenLauncherModalOpen: true
-    })
-  }
-
-  handleClose = () => {
-    this.setState({
-      screenLauncherModalOpen: false
-    })
-  }
 
   handleControlsByKeyboard = (e) => {
     switch(e.keyCode) {
@@ -57,7 +23,7 @@ class MenuComponent extends React.Component {
         this.props.prev();
         return 0;
       case 40: // Arrow down
-        this.handleOpen();
+        this.screenLauncherRef.getWrappedInstance().open();
         return 0;
       case 32: // Space
         if(this.props.isPlaying) {
@@ -70,32 +36,17 @@ class MenuComponent extends React.Component {
     }
   }
 
-
   render() {
-    const {formatMessage} = this.props.intl;
-
-    const actions = [
-      <FlatButton
-        label={formatMessage(messages.closeButton)}
-        primary={true}
-        onTouchTap={this.handleClose}
-      />
-    ];
-
     return (
       <div className='menu-component'>
 
         <KeyBinding onKey={ (e) => { this.handleControlsByKeyboard(e) } } />
 
-        <div className='main-menu-fab-container'>
-          <FloatingActionButton
-            disabled={this.props.slides.length === 0}
-            mini={true}
-            secondary={true}
-            onTouchTap={this.handleOpen} >
-            <ContentLowPriority />
-          </FloatingActionButton>
-        </div>
+        <ScreenLauncherComponent
+          ref={(c) => this.screenLauncherRef = c}
+          slides={this.props.slides}
+          goto={this.props.goto}
+        />
 
         <div className='main-menu-fab-container'>
           <FloatingActionButton
@@ -144,21 +95,6 @@ class MenuComponent extends React.Component {
           />
         </div>
 
-        <Dialog
-          title={formatMessage(messages.title)}
-          actions={actions}
-          modal={false}
-          open={this.state.screenLauncherModalOpen}
-          onRequestClose={this.handleClose}
-          autoScrollBodyContent={true}
-        >
-          <ScreenLauncher
-            goto={this.props.goto}
-            handleClose={this.handleClose}
-            slides={this.props.slides}
-          />
-        </Dialog>
-
       </div>
     );
   }
@@ -175,4 +111,4 @@ MenuComponent.propTypes = {
   goto: React.PropTypes.func.isRequired
 };
 
-export default injectIntl(MenuComponent);
+export default MenuComponent;
